+++ src/Spice86/DependencyInjection/TopLevelContainer.cs
@@ -1,0 +1,15 @@
+namespace Spice86.DependencyInjection;
+
+using Spice86.Logging;
+using Spice86.Shared.Interfaces;
+
+using StrongInject;
+
+/// <summary>
+/// Provides top-level services
+/// </summary>
+[Register(typeof(LoggerPropertyBag), Scope.SingleInstance, typeof(ILoggerPropertyBag))]
+[Register(typeof(LoggerService), Scope.SingleInstance, typeof(ILoggerService))]
+[Register(typeof(Startup))]
+public partial class TopLevelContainer : IContainer<Startup> {
+}
