+++ src/Spice86/Spice86.csproj
@@ -33,16 +33,16 @@
 		<ContinuousIntegrationBuild Condition="'$(GITHUB_ACTIONS)' == 'true'">true</ContinuousIntegrationBuild>
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
 		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="All" />
 		<PackageReference Include="DotNet.ReproducibleBuilds" Version="1.1.1">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="OxyPlot.Avalonia" Version="2.1.0" />
 	</ItemGroup>
 
 	<ItemGroup>
+		<PackageReference Include="StrongInject" Version="1.4.4" />
+		<PackageReference Include="OxyPlot.Avalonia" Version="2.1.0" />
 		<PackageReference Include="Ben.Demystifier" Version="0.4.1" />
 		<PackageReference Include="MessageBox.Avalonia" Version="1.7.1" />
 		<PackageReference Include="CommunityToolkit.Mvvm" Version="8.2.0" />
